+++ pkg/tracing/init.go
@@ -108,6 +108,7 @@ func initRoutes(ctx context.Context, app *bunapp.App, sp *SpanProcessor) {
 			g.GET("/spans", spanHandler.ListSpans)
 			g.GET("/percentiles", spanHandler.Percentiles)
 			g.GET("/group-stats", spanHandler.GroupStats)
+			g.GET("/timeseries", spanHandler.Timeseries)
 		})
 
 	api.